Move HyperBosses to its own file
